requirements.txt
pymongo~=3.11.4
python-dateutil~=2.8.1
watchdog~=2.1.1
Flask~=2.0.1
pytz~=2021.1
python-geoip-python3~=1.3
python-geoip~=1.2
natsort~=7.1.1
tzlocal~=2.1
netifaces~=0.11.0
xmpppy~=0.6.2
humanfriendly~=9.2
geoip2~=4.4.0
requests~=2.26.0
pydbus~=0.6.0
